+++ apps/validator/src/main/scala/org/lfdecentralizedtrust/splice/validator/automation/ValidatorAutomationService.scala
@@ -42,6 +42,7 @@ import io.opentelemetry.api.trace.Tracer
 import monocle.Monocle.toAppliedFocusOps
 import org.apache.pekko.stream.Materializer
 import org.lfdecentralizedtrust.splice.store.AppStoreWithIngestion.SpliceLedgerConnectionPriority
+import org.lfdecentralizedtrust.splice.validator.metrics.ValidatorAppMetrics
 
 import java.nio.file.Path
 import scala.concurrent.ExecutionContextExecutor
@@ -77,6 +78,7 @@ class ValidatorAutomationService(
     maxVettingDelay: NonNegativeFiniteDuration,
     params: SpliceParametersConfig,
     latestPackagesOnly: Boolean,
+    metrics: ValidatorAppMetrics,
     override protected val loggerFactory: NamedLoggerFactory,
 )(implicit
     ec: ExecutionContextExecutor,
@@ -97,6 +99,19 @@ class ValidatorAutomationService(
   override def companion
       : org.lfdecentralizedtrust.splice.validator.automation.ValidatorAutomationService.type =
     ValidatorAutomationService
+
+  automationConfig.topologyMetricsPollingInterval.foreach(topologyPollingInterval =>
+    registerTrigger(
+      new TopologyMetricsTrigger(
+        triggerContext
+          .focus(_.config.pollingInterval)
+          .replace(topologyPollingInterval),
+        metrics,
+        scanConnection,
+        participantAdminConnection,
+      )
+    )
+  )
 
   walletManagerOpt.foreach { walletManager =>
     registerTrigger(